Iterating
Switched to SetArrayResult FALSE, which means the sphinx results are an
array where the array key is the post_id.

Cascaded changes after switching SetArrayResult FALSE eliminated
iterators in two other methods, possibly improving performance.

Adding lists of both Sphinx and MySQL post ids to the debug output.

Doing a console.log on the debug output, for more convenient viewing.

Added a line to the debug HTML to state if the Sphinx and MySQL query
results match.

This needs more cleanup, just doing a WIP commit.